Sample Dungeon Collision Scene
Created a sample dungeon scene with a controlable character for anyone who is interested in messing around.

A simple next step would be to update the collision boxes of the dungeon walls to allow the character to fit up into the smaller regions of the room. (Remove some of the collision boxes from the upperside of dungeon walls.)